Changing queries and increasing range of numbers

+++ database/sql/create_tables.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS range_central (
   rangeid INT NOT NULL,
-  lower INT NOT NULL,
-  higher INT NOT NULL,
+  lower BIGINT NOT NULL,
+  higher BIGINT NOT NULL,
   node varchar(200),
   created TIMESTAMP,
   expired INT NOT NULL,
